+++ lib/baby_squeel.rb
@@ -5,29 +5,6 @@ require 'baby_squeel/version'
 require 'baby_squeel/errors'
 
 module BabySqueel
-  class << self
-    # Configures BabySqueel using the given block
-    def configure
-      yield self
-    end
-
-    # Get a BabySqueel table instance.
-    #
-    # ==== Examples
-    #     BabySqueel[Post]
-    #     BabySqueel[:posts]
-    #     BabySqueel[Post.arel_table]
-    #
-    def [](thing, **kwargs)
-      if thing.respond_to?(:model_name)
-        Relation.new(thing)
-      elsif thing.kind_of?(Arel::Table)
-        Table.new(thing)
-      else
-        Table.new(Arel::Table.new(thing, **kwargs))
-      end
-    end
-  end
 end
 
 ActiveSupport.on_load :active_record do