+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- remove BabySqueel.configure and BabySqueel[]
 - remove BabySqueel::ActiveRecord::Calculations (plucking, averaging, counting, maximizing, minimizing, summing)
 - remove reordering, ordering, grouping, when_having
 - remove Operators::Grouping, Operators::Matching, Operators::Comparison